Use abstract class to load rankings from resources
